requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
boto3==1.13.11
botocore==1.16.11
celery==4.4.2
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.0
django-filter==2.2.0
django-render-block==0.6
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
kombu==4.6.8
Markdown==3.2.1
Pillow==7.1.1
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.2
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
